Move description based heuristic into own module
